Improving how current handles gems with multiple versions installed
Current() didn't deal with multiple versions of the same gem properly.
It would see current as '1.2.3a.b.c' and try to install 1.2.3 again.
Also comma-seperate the values when displaying them if there are multiple
versions.  Moved this logic out of list() because it really has to do with
picking what the current value is.