Change how we shell out for gem to be more consistent with how RBEnv works in the shell
RBEnv gem installs were failing for me.  I changed the module to stop using
an explicit path to gem and set the search path in the environment the same
way that RBEnv does things in an interactive shell and that worked.  I also
changed resource[:rbenv] to pass the RBEnv root, rather than the path to the
desired ruby version because the name was inaccurate and what I really
needed was the RBEnv root anyway.