Added caching and fixed version splitting.
Added a little caching so that it wouldn't pull the gem list more than once
for each gem.  Also changed how the version list from gem gets split because
for some reason, nokogiri under JRuby was reporting a version of 'x.y.z
java' when I was trying to install x.y.z, which made it keep trying to
reinstall it.